apps/polymarket-spike-bot-v1/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
web3==5.31.3
py-clob-client==0.1.0
py-clob-client==0.1.0
halo
fastapi
uvicorn
python-multipart

